Convert Config to kapi.List{} and fix Template to use runtime.Object{}
