+++ pkg/config/api/types.go
@@ -1,18 +1,7 @@
 package api
 
-import (
-	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
-)
+import kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
 
-// Config contains a set of Kubernetes resources to be applied.
-// TODO: Unify with Kubernetes Config
-//       https://github.com/GoogleCloudPlatform/kubernetes/pull/1007
-type Config struct {
-	kapi.TypeMeta   `json:",inline"`
-	kapi.ObjectMeta `json:"metadata,omitempty"`
-
-	// Required: Items is an array of Kubernetes resources of Service,
-	// Pod and/or ReplicationController kind.
-	Items []runtime.RawExtension `json:"items"`
-}
+// Config implements the Kubernetes api.List
+// DEPRECATED: The version v1beta2 should use api.List instead of Config
+type Config kapi.List
