+++ pkg/config/api/v1beta1/types.go
@@ -5,9 +5,8 @@ import (
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
 )
 
-// Config contains a set of Kubernetes resources to be applied.
-// TODO: Unify with Kubernetes Config
-//       https://github.com/GoogleCloudPlatform/kubernetes/pull/1007
+// Config contains a list of Kubernetes resources to be applied.
+// DEPRECATED: All clients should use Kubernetes api.List instead of Config.
 type Config struct {
 	kapi.TypeMeta   `json:",inline"`
 	kapi.ObjectMeta `json:"metadata,omitempty"`
