+++ pkg/template/labels_test.go
@@ -1,116 +1,14 @@
-package config
+package template
 
 import (
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
 	"reflect"
 	"testing"
 
 	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
-	klatest "github.com/GoogleCloudPlatform/kubernetes/pkg/api/latest"
 	kmeta "github.com/GoogleCloudPlatform/kubernetes/pkg/api/meta"
-	kclient "github.com/GoogleCloudPlatform/kubernetes/pkg/client"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/kubectl/resource"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
-
-	clientapi "github.com/openshift/origin/pkg/cmd/client/api"
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
 )
-
-func TestApplyInvalidConfig(t *testing.T) {
-	clients := clientapi.ClientMappings{
-		"InvalidClientMapping": {"InvalidClientResource", nil, nil},
-	}
-	clientFunc := func(m *kmeta.RESTMapping) (*resource.Helper, error) {
-		mapping, ok := clients[m.Resource]
-		if !ok {
-			return nil, fmt.Errorf("Unable to provide REST client for %v", m.Resource)
-		}
-		return resource.NewHelper(mapping.Client, m), nil
-	}
-	invalidConfigs := []string{
-		`{}`,
-		`{ "foo": "bar" }`,
-		`{ "items": null }`,
-		`{ "items": "bar" }`,
-		`{ "items": [ null ] }`,
-		`{ "items": [ { "foo": "bar" } ] }`,
-		`{ "items": [ { "kind": "", "apiVersion": "v1beta1" } ] }`,
-		`{ "items": [ { "kind": "UnknownResource", "apiVersion": "v1beta1" } ] }`,
-		`{ "items": [ { "kind": "InvalidClientResource", "apiVersion": "v1beta1" } ] }`,
-	}
-	for i, invalidConfig := range invalidConfigs {
-		result, err := Apply(kapi.NamespaceDefault, []byte(invalidConfig), clientFunc)
-
-		if i <= 3 && err == nil {
-			t.Errorf("Expected error while applying invalid Config '%v', result: %v", invalidConfigs[i], result)
-		}
-
-		for _, itemResult := range result {
-			if len(itemResult.Errors) > 0 {
-				t.Errorf("Expected error while applying invalid Config '%v'", invalidConfigs[i])
-			}
-		}
-	}
-}
-
-func TestApplySendsData(t *testing.T) {
-	received := make(chan bool, 1)
-	fakeServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		received <- true
-		validUrl := "/api/v1beta1/ns/" + kapi.NamespaceDefault + "/pods"
-		if r.RequestURI != validUrl {
-			t.Errorf("Unexpected RESTClient RequestURI. Expected: %v, got: %v.", validUrl, r.RequestURI)
-		}
-	}))
-
-	uri, _ := url.Parse(fakeServer.URL + "/api/v1beta1")
-	fakeClient := kclient.NewRESTClient(uri, "v1beta1", kapi.Codec, false)
-	clients := clientapi.ClientMappings{
-		"pods": {"Pod", fakeClient, kapi.Codec},
-	}
-	clientFunc := func(m *kmeta.RESTMapping) (*resource.Helper, error) {
-		mapping, ok := clients[m.Resource]
-		if !ok {
-			return nil, fmt.Errorf("Unable to provide REST client for %v", m.Resource)
-		}
-		return resource.NewHelper(mapping.Client, m), nil
-	}
-	config := `{ "apiVersion": "v1beta1", "kind": "Config", "metadata" : { "name": "test-config" }, "items": [ { "kind": "Pod", "apiVersion": "v1beta1", "metadata": { "name": "FakePod" } } ] }`
-	result, err := Apply(kapi.NamespaceDefault, []byte(config), clientFunc)
-
-	if err != nil || result == nil {
-		t.Errorf("Unexpected error while applying valid Config '%v', result: %v, error: %v", config, result, err)
-	}
-
-	for _, itemResult := range result {
-		if len(itemResult.Errors) > 0 {
-			t.Errorf("Unexpected error while applying valid Config '%v': %+v", config, itemResult.Errors)
-		}
-	}
-
-	// <-received
-}
-
-func ExampleApply() {
-	kubeClient, _ := kclient.New(&kclient.Config{Host: "127.0.0.1"})
-	testClientMappings := clientapi.ClientMappings{
-		"pods":     {"Pod", kubeClient.RESTClient, klatest.Codec},
-		"services": {"Service", kubeClient.RESTClient, klatest.Codec},
-	}
-	clientFunc := func(m *kmeta.RESTMapping) (*resource.Helper, error) {
-		mapping, ok := testClientMappings[m.Resource]
-		if !ok {
-			return nil, fmt.Errorf("Unable to provide REST client for %v", m.Resource)
-		}
-		return resource.NewHelper(mapping.Client, m), nil
-	}
-	data, _ := ioutil.ReadFile("../../examples/sample-app/docker-registry-config.json")
-	Apply(kapi.NamespaceDefault, data, clientFunc)
-}
 
 type FakeLabelsResource struct {
 	kapi.TypeMeta   `json:",inline"`
@@ -319,6 +217,116 @@ func TestAddConfigLabels(t *testing.T) {
 			if e, a := test.expectedLabels, objType.Template.ControllerTemplate.Template.Labels; !reflect.DeepEqual(e, a) {
 				t.Errorf("Unexpected labels on testCase[%v]. Expected: %#v, got: %#v.", i, e, a)
 			}
+		}
+	}
+}
+
+func TestMergeInto(t *testing.T) {
+	var nilMap map[int]int
+
+	testCases := []struct {
+		dst      interface{}
+		src      interface{}
+		flags    int
+		err      bool
+		expected interface{}
+	}{
+		{ // [0] Can't merge into nil
+			dst:      nil,
+			src:      map[int]int{},
+			flags:    0,
+			err:      true,
+			expected: nil,
+		},
+		{ // [1] Can't merge untyped nil into an empty map
+			dst:      map[int]int{},
+			src:      nil,
+			flags:    0,
+			err:      true,
+			expected: map[int]int{},
+		},
+		{ // [2] Merge nil map into an empty map
+			dst:      map[int]int{},
+			src:      nilMap,
+			flags:    0,
+			err:      false,
+			expected: map[int]int{},
+		},
+		{ // [3] Can't merge into nil map
+			dst:      nilMap,
+			src:      map[int]int{},
+			flags:    0,
+			err:      true,
+			expected: nilMap,
+		},
+		{ // [4] Can't merge into pointer
+			dst:      &nilMap,
+			src:      map[int]int{},
+			flags:    0,
+			err:      true,
+			expected: &nilMap,
+		},
+		{ // [5] Test empty maps
+			dst:      map[int]int{},
+			src:      map[int]int{},
+			flags:    0,
+			err:      false,
+			expected: map[int]int{},
+		},
+		{ // [6] Test dst + src => expected
+			dst:      map[int]byte{0: 0, 1: 1},
+			src:      map[int]byte{2: 2, 3: 3},
+			flags:    0,
+			err:      false,
+			expected: map[int]byte{0: 0, 1: 1, 2: 2, 3: 3},
+		},
+		{ // [7] Test dst + src => expected, do not overwrite dst
+			dst:      map[string]string{"foo": "bar"},
+			src:      map[string]string{"foo": ""},
+			flags:    0,
+			err:      false,
+			expected: map[string]string{"foo": "bar"},
+		},
+		{ // [8] Test dst + src => expected, overwrite dst
+			dst:      map[string]string{"foo": "bar"},
+			src:      map[string]string{"foo": ""},
+			flags:    OverwriteExistingDstKey,
+			err:      false,
+			expected: map[string]string{"foo": ""},
+		},
+		{ // [9] Test dst + src => expected, error on existing key value
+			dst:      map[string]string{"foo": "bar"},
+			src:      map[string]string{"foo": "bar"},
+			flags:    ErrorOnExistingDstKey | OverwriteExistingDstKey,
+			err:      true,
+			expected: map[string]string{"foo": "bar"},
+		},
+		{ // [10] Test dst + src => expected, do not error on same key value
+			dst:      map[string]string{"foo": "bar"},
+			src:      map[string]string{"foo": "bar"},
+			flags:    ErrorOnDifferentDstKeyValue | OverwriteExistingDstKey,
+			err:      false,
+			expected: map[string]string{"foo": "bar"},
+		},
+		{ // [11] Test dst + src => expected, error on different key value
+			dst:      map[string]string{"foo": "bar"},
+			src:      map[string]string{"foo": ""},
+			flags:    ErrorOnDifferentDstKeyValue | OverwriteExistingDstKey,
+			err:      true,
+			expected: map[string]string{"foo": "bar"},
+		},
+	}
+
+	for i, test := range testCases {
+		err := MergeInto(test.dst, test.src, test.flags)
+		if err != nil && !test.err {
+			t.Errorf("Unexpected error while merging maps on testCase[%v]: %v.", i, err)
+		} else if err == nil && test.err {
+			t.Errorf("Unexpected non-error while merging maps on testCase[%v].", i)
+		}
+
+		if !reflect.DeepEqual(test.dst, test.expected) {
+			t.Errorf("Unexpected map on testCase[%v]. Expected: %#v, got: %#v.", i, test.expected, test.dst)
 		}
 	}
 }
