--- pkg/util/mergemap.go
@@ -1,66 +1,0 @@
-package util
-
-import (
-	"fmt"
-	"reflect"
-)
-
-// MergeInto flags
-const (
-	OverwriteExistingDstKey = 1 << iota
-	ErrorOnExistingDstKey
-	ErrorOnDifferentDstKeyValue
-)
-
-// MergeInto merges items from a src map into a dst map.
-// Returns an error when the maps are not of the same type.
-// Flags:
-// - ErrorOnExistingDstKey
-//     When set: Return an error if any of the dst keys is already set.
-// - ErrorOnDifferentDstKeyValue
-//     When set: Return an error if any of the dst keys is already set
-//               to a different value than src key.
-// - OverwriteDstKey
-//     When set: Overwrite existing dst key value with src key value.
-func MergeInto(dst, src interface{}, flags int) error {
-	dstVal := reflect.ValueOf(dst)
-	srcVal := reflect.ValueOf(src)
-
-	if dstVal.Kind() != reflect.Map {
-		return fmt.Errorf("dst is not a valid map: %v", dstVal.Kind())
-	}
-	if srcVal.Kind() != reflect.Map {
-		return fmt.Errorf("src is not a valid map: %v", srcVal.Kind())
-	}
-	if dstTyp, srcTyp := dstVal.Type(), srcVal.Type(); !dstTyp.AssignableTo(srcTyp) {
-		return fmt.Errorf("type mismatch, can't assign '%v' to '%v'", srcTyp, dstTyp)
-	}
-
-	if dstVal.IsNil() {
-		return fmt.Errorf("dst value is nil")
-	}
-	if srcVal.IsNil() {
-		// Nothing to merge
-		return nil
-	}
-
-	for _, k := range srcVal.MapKeys() {
-		if dstVal.MapIndex(k).IsValid() {
-			if flags&ErrorOnExistingDstKey != 0 {
-				return fmt.Errorf("dst key already set (ErrorOnExistingDstKey=1), '%v'='%v'", k, dstVal.MapIndex(k))
-			}
-			if dstVal.MapIndex(k).String() != srcVal.MapIndex(k).String() {
-				if flags&ErrorOnDifferentDstKeyValue != 0 {
-					return fmt.Errorf("dst key already set to a different value (ErrorOnDifferentDstKeyValue=1), '%v'='%v'", k, dstVal.MapIndex(k))
-				}
-				if flags&OverwriteExistingDstKey != 0 {
-					dstVal.SetMapIndex(k, srcVal.MapIndex(k))
-				}
-			}
-		} else {
-			dstVal.SetMapIndex(k, srcVal.MapIndex(k))
-		}
-	}
-
-	return nil
-}
